Reorganizing parser and simplifying lexer
This commit updates the parser to remove a few layers of abstractions
by relying more more on "operators" rather than Tokens when dealing
with precedence parsing. This change removes a number of methods from
Token that are no longer necessary, including tracking token precedence
in the parser's Operator enum rather than the Token enum.

An effect of this change is a significant speedup in parsing. For
example, bench_deep_projection_104 has improved from 101,738 ns/iter to
56,132 ns/iter.