Creating a Variable type and using more Rc
This commit updates the library to use its own variable type. This is
necessary in order to support expression references, and has the added
benefit of making everything in the variable type a reference counted,
Rc, pointer -- avoiding copies and ensuring that values are dropped when
their ref count is equal to 0.