Code generating compliance tests and benches
This commit adds support for code generating compliance tests as regular
test cases. This has the added benefit of being able to see exactly
which case failed, and uses the built-in support for displaying
success/fail metrics.

This also adds support for code generating benchmarks from compliance
tests that have a value of "bench" set to "parse" or "full" where
"parse" means only benchmark parsing an expression, and "full" means
benchmark parsing and interpreting. The generated benchmarks are then
run when `cargo bench` is invoked, showing the comment or slugified
expression as part of the derived benchmark name.

Closes #10