Adding serde serialization/deserialization.
This commit adds serialization and deserialization to JMESPath
Variables. This allows the value needed by JMESPath's runtime
to be constructed directly from a JSON string without an
intermediate Value type and written to a string/writer without
an intermediate Value. It also allows for a much broader
ToJMESPath trait as we can leverage serde's serialization
traits.

Closes #4