Updating functions to fn with dispatcher.
This commit updates the function system to no longer be a
HashMap<String, FnBox>, but rather is now a FnDispatcher that handles
the dispatching inside of the trait implementation. This allows for a
more composable function system (for example you could decorate the
existing built-in dispatcher with custom functions), and provides about
a 30% performance improvement for the interpreter because a large
portion of the time was spent to building up and boxing the entire set
of possible functions in the hash map. Now we are instead just calling
actual functions directly.