Expression is now Sync.
This commit updates Expression to use a serde_json::Value as the literal
value instad of an RcVar. This has the benefit that Expression is now
Sync and therefore can be used in things like lazy_static!, but it does
require that a value is converted to a Variable each time a literal
value is referenced while evaluating an expression. Most literals tend
to be rather small and typically scalars, so this should not cause a
noticeable performance issue in practice.