Moving back to recursive parser, fixing tests
This commit moves back to the much simpler recursive pratt parser. There
were a number of failing tests with the trampolining implementation that
were not detected because there was a bug in the test runner as well.
The test runner has now been updated to perform equality checks on all
result values, not just strings as it did previously. This commit also
adds much less strict float equality checks to be able to pass the
compliance tests and compare numbers similar to other JP
implementations.

Closes #16