Problem: serde_json 1.0.45 breaks jmespath.rs
It fails to compile:

```
error[E0603]: enum `ErrorCode` is private
 --> src/variable.rs:9:32
  |
9 | use serde_json::error::{Error, ErrorCode};
  |                                ^^^^^^^^^

error[E0624]: method `syntax` is private
    --> src/variable.rs:1205:29
     |
1205 |             _ => return Err(Error::syntax(ErrorCode::KeyMustBeAString, 0, 0)),
```

This happens because serde_json made `Error::syntax` and `ErrorCode`
available only to `serde_json` itself.

Solution: make serde_json trigger an error we need

It's an ugly kludge but at least it works :) It creates a HashMap with a
non-string key and passes this off to serde_json. It triggers the error
we need.

This is unlikely to be a great solution long term, but at least it makes
jmespath.rs compile without pinning it to 1.0.44 forever.